genres/requirements.txt
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
apsw==3.8.2.post1
apt-xapian-index==0.45
asp==0.1.3.1
avro==1.7.7
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
beautifulsoup4==4.2.1
boto==2.20.1
brewer2mpl==1.4.1
Brlapi==0.6.1
bzr==2.7.0.dev1
ccsm==0.9.11.3
certifi==2016.2.28
cgen==2015.1.1
chardet==2.0.1
CherryPy==3.2.2
codepy==2013.1.2
colorama==0.2.5
command-not-found==0.3
compizconfig-python==0.9.11.3
configglue==1.1.2
configobj==4.7.2
configparser==3.3.0.post2
cssselect==0.9.1
cssutils==0.9.10
cvxopt==1.1.4
cycler==0.9.0
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
discover==0.4.0
dnspython==1.11.1
duplicity==0.6.23
entrypoints==0.2
enum34==1.1.3
eyeD3==0.7.9
feedparser==5.1.3
fluendo-eula==0.6
functools32==3.2.3.post2
futures==3.0.5
ggplot==0.6.8
GnuPGInterface==0.3.2
h5py==2.2.1
html5lib==0.999
httplib2==0.8
ipykernel==4.3.1
ipyparallel==5.0.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
joblib==0.7.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
louis==2.5.3
lxml==3.3.3
Mako==0.9.1
Markdown==2.4
MarkupSafe==0.18
matplotlib==1.5.0
mechanize==0.2.5
mistune==0.7.2
mlpy==0.1.0
mock==1.0.1
mutagen==1.22
nbconvert==4.2.0
nbformat==4.0.1
netifaces==0.8
nose==1.3.1
notebook==4.2.0
numexpr==2.2.2
numpy==1.8.1
oauth==1.0.1
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
openpyxl==1.7.0
PAM==0.4.2
panda==0.3.1
pandas==0.13.1
paramiko==1.10.1
pathlib2==2.1.0
patsy==0.2.1
pexpect==3.1
pickleshare==0.7.2
Pillow==2.3.0
piston-mini-client==0.7.5
protobuf==2.5.0
ptyprocess==0.5.1
pycasp==0.1
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
Pygments==1.6
pygobject==3.12.0
pygpgme==0.3
pyinotify==0.9.4
pyOpenSSL==0.13
pyparsing==2.0.1
pyserial==2.6
pysmbc==1.0.14.1
python-apt===0.9.3.5ubuntu2
python-cloudfiles==1.7.10
python-dateutil==1.5
python-debian===0.1.21-nmu2ubuntu2
python-distutils-extra==2.38
python-mpd==0.3.0
pytools==2015.1.6
pytz==2012rc0
pyxdg==0.25
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.2.1
reportlab==3.0
repoze.lru==0.6
requests==2.2.1
rhythmbox-ubuntuone==4.2.0
Routes==2.0
scikit-learn==0.17
scikits.audiolab==0.11.0
scipy==0.13.3
SecretStorage==2.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.3.1
singledispatch==3.4.0.3
six==1.10.0
software-center-aptd-plugins==0.0.0
speech-features==0.1
ssh-import-id==3.21
statsmodels==0.6.1
system-service==0.1.6
tables==3.1.1
terminado==0.6
tornado==4.3
traitlets==4.2.1
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
unittest2==0.5.1
unity-lens-photos==1.0
urllib3==1.7.1
uTidylib==0.2
vboxapi==1.0
virtkey==0.63.0
virtualenv==15.0.1
wadllib==1.3.2
WebOb==1.3.1
xdiagnose===3.6.3build2
xkit==0.0.0
xlrd==0.9.2
xlwt==0.7.5
zope.interface==4.0.5
